Project 1
The inconvenience is regretted because few files are jumbled and may not be known their uses. please make sure you change the name and the port of mysql to your current port as I have changed mine's to port to 3307 unlike the other computers which use port 3306. If you hav the problem with the ports, Goto Task manager and end the sources which use that source(eg: VM ware usually uses the port 3306 to communicate with the computer) than itself you can run your mysql there.